+++ CHANGELOG.md
@@ -21,7 +21,7 @@
 * Fixed bug with removing builds that have been deleted in Raven Colonial
 * Made debug logging only in dev_mode
 * Fixed situation where a site might not get associated with the appropriate projectid
-
+* Fix for missing rare surface station update situation
 
 ## v5.2.0-b2 - 2025-10-12